+++ Sources/UIComponent/Core/Model/RenderNode/RenderNodeModifiers.swift
@@ -2,44 +2,6 @@
 
 import UIKit
 
-public protocol RenderNodeWrapper: RenderNode {
-    associatedtype Content: RenderNode
-    var content: Content { get }
-}
-
-extension RenderNodeWrapper {
-    public var id: String? {
-        content.id
-    }
-    public var animator: Animator? {
-        content.animator
-    }
-    public var reuseStrategy: ReuseStrategy {
-        content.reuseStrategy
-    }
-    public var shouldRender: Bool {
-        content.shouldRender
-    }
-    public var size: CGSize {
-        content.size
-    }
-    public var positions: [CGPoint] {
-        content.positions
-    }
-    public var children: [any RenderNode] {
-        content.children
-    }
-    public func visibleIndexes(in frame: CGRect) -> IndexSet {
-        content.visibleIndexes(in: frame)
-    }
-    public func updateView(_ view: Content.View) {
-        content.updateView(view)
-    }
-    public func makeView() -> Content.View {
-        content.makeView()
-    }
-}
-
 public struct UpdateRenderNode<Content: RenderNode>: RenderNodeWrapper {
     public let content: Content
     public let update: (Content.View) -> Void
